유저 프로필 검수 api 구현

+++ src/main/java/com/lingo/lingoproject/admin/presentation/dto/ProfileReviewListResponseDto.java
@@ -8,8 +8,8 @@ public record ProfileReviewListResponseDto(
     List<ProfileReviewResponseDto> content,
     int page,
     int size,
-    int totalElements,
-    int totalPages
+    long totalElements,
+    long totalPages
 ) {
 
 }
